move VpcSubnet views and params into nexus
